pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proofpoint_itm"
version = "0.9.0"
description = "Proofpoint ITM API client library"
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = ["requests", "pydantic"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
authors = [
    { name = "Mike Olden", email = "michael.olden@oldendigital.com" },
]
license = "Apache-2.0"
license-files = ["LICENSE"]
keywords = [
    "Proofpoint ITM",
    "Proofpoint",
    "ITM",
    "ObserveIT"
]

[project.urls]
"Homepage" = "https://github.com/drizzo-tech/proofpoint_itm"
"Repository" = "https://github.com/drizzo-tech/proofpoint_itm"
"Documentation" = "https://proofpoint-itm.readthedocs.io/en/latest/index.html"
"Issues" = "https://github.com/drizzo-tech/proofpoint_itm/issues"

[project.optional-dependencies]
build = [
    "setuptools",
    "wheel",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-xdist>=3.3.0",
    "responses>=0.23.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-xdist>=3.3.0",
    "responses>=0.23.0",
    "python-dotenv>=1.0.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
    "mypy>=1.4.0",
]